+++ Editor/SettingsEditor.cs
@@ -1,14 +1,14 @@
-﻿//#define ENABLE_LEGACY_PREFERENCE
+﻿#define ENABLE_HANANOKI_SETTINGS
 
 using UnityEditor;
 using UnityEngine;
-using Hananoki.Extensions;
-using Hananoki.SharedModule;
+using HananokiEditor.Extensions;
+using HananokiEditor.SharedModule;
 
-using E = Hananoki.CustomHierarchy.SettingsEditor;
-using SS = Hananoki.SharedModule.S;
+using E = HananokiEditor.CustomHierarchy.SettingsEditor;
+using SS = HananokiEditor.SharedModule.S;
 
-namespace Hananoki.CustomHierarchy {
+namespace HananokiEditor.CustomHierarchy {
 	[System.Serializable]
 	public class SettingsEditor {
 		public bool Enable = true;
@@ -178,13 +178,12 @@ namespace Hananoki.CustomHierarchy {
 
 
 #if ENABLE_HANANOKI_SETTINGS
-	[SettingsClass]
 	public class SettingsEvent {
-		[SettingsMethod]
+		[HananokiSettingsRegister]
 		public static SettingsItem RegisterSettings() {
 			return new SettingsItem() {
 				//mode = 1,
-				displayName = Package.name,
+				displayName = Package.nameNicify,
 				version = Package.version,
 				gui = SettingsEditorWindow.DrawGUI,
 			};